+++ noodles-bgzf/src/gzi/async/reader.rs
@@ -65,7 +65,7 @@ where
                 io::ErrorKind::InvalidData,
                 "unexpected trailing data",
             )),
-            Err(ref e) if e.kind() == io::ErrorKind::UnexpectedEof => Ok(offsets),
+            Err(ref e) if e.kind() == io::ErrorKind::UnexpectedEof => Ok(Index::from(offsets)),
             Err(e) => Err(e),
         }
     }
@@ -89,7 +89,7 @@ mod tests {
 
         assert_eq!(
             reader.read_index().await?,
-            vec![(0, 0), (4668, 21294), (23810, 86529)]
+            Index::from(vec![(0, 0), (4668, 21294), (23810, 86529)])
         );
 
         Ok(())
@@ -101,7 +101,7 @@ mod tests {
 
         let mut reader = Reader::new(&data[..]);
         let index = reader.read_index().await?;
-        assert_eq!(index, [(0, 0)]);
+        assert_eq!(index, Index::from(vec![(0, 0)]));
 
         Ok(())
     }