+++ noodles-bgzf/src/reader.rs
@@ -190,11 +190,11 @@ where
     ///
     /// ```
     /// # use std::io;
-    /// use noodles_bgzf as bgzf;
+    /// use noodles_bgzf::{self as bgzf, gzi};
     ///
     /// let mut reader = bgzf::Reader::new(io::empty());
     ///
-    /// let index = vec![(0, 0)];
+    /// let index = gzi::Index::from(vec![(0, 0)]);
     /// reader.seek_by_uncompressed_position(&index, 0)?;
     /// # Ok::<_, io::Error>(())
     /// ```
@@ -203,6 +203,8 @@ where
         index: &gzi::Index,
         pos: u64,
     ) -> io::Result<u64> {
+        let index = index.as_ref();
+
         assert!(!index.is_empty());
 
         let i = index.partition_point(|r| r.1 <= pos);
@@ -404,7 +406,7 @@ mod tests {
             0x02, 0x00, 0x1b, 0x00, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
         ];
 
-        let index = vec![(0, 0), (35, 7)];
+        let index = gzi::Index::from(vec![(0, 0), (35, 7)]);
 
         let mut reader = Reader::new(Cursor::new(&data));